+++ pdu_bind_resp.go
@@ -116,8 +116,8 @@ func (s *BindResp) writeTLVFields() []byte {
 func (s *BindResp) Writer() []byte {
 	b := append(s.writeFields(), s.writeTLVFields()...)
 	h := packUi32(uint32(len(b) + 16))
-	h = append(h, packUi32(s.Header.Id)...)
-	h = append(h, packUi32(s.Header.Status)...)
+	h = append(h, packUi32(uint32(s.Header.Id))...)
+	h = append(h, packUi32(uint32(s.Header.Status))...)
 	h = append(h, packUi32(s.Header.Sequence)...)
 	return append(h, b...)
 }
